Move dynamic object into its own package for cleanliness.
