+++ pkg/dynamic/object.go
@@ -1,4 +1,4 @@
-package watcher
+package dynamic
 
 import (
 	"fmt"
