+++ pkg/dynamic/object_test.go
@@ -1,4 +1,4 @@
-package watcher
+package dynamic
 
 import (
 	"reflect"
